advanced-mate-menu: do not package the configuration on vojtux-settings, enable it only on live image
The reason is that if we would like to package the configuration as a installable package, we would need to consider the case when someone installs the package on a system with existing user Mate configuration.
I do not think that it is possible to append dconf configuration of Mate panel to existing one.
therefore, there is a chance we would broke someone's existing Mat panel configuration.
Let's solve this rather with documentation.